reconsolidate source files
