--- SDKDemo/Plugins/HathoraSDK/Source/HathoraSDK/Public/RegionalPings.h
@@ -1,13 +1,0 @@
-// Copyright 2023 Hathora, Inc.
-#pragma once
-
-#include "RegionalPings.generated.h"
-
-USTRUCT(BlueprintType)
-struct HATHORASDK_API FRegionalPings
-{
-	GENERATED_BODY()
-
-	UPROPERTY(BlueprintReadOnly)
-	TMap<FString, int32> PingMap;
-};
